PURGE THE IMPURE
