+++ signatures/src/main/scala/tsec/signature/imports/JCASigInterpreter.scala
@@ -4,7 +4,7 @@ import java.security.Signature
 import cats.effect.Sync
 import tsec.signature.core._
 
-sealed abstract class JCASigInterpreterPure[F[_], A](implicit M: Sync[F], signatureAlgorithm: SigAlgoTag[A])
+sealed abstract class JCASigInterpreter[F[_], A](implicit M: Sync[F], signatureAlgorithm: SigAlgoTag[A])
     extends SignatureAlgebra[F, A] {
   type S     = Signature
   type PubK  = SigPublicKey[A]
@@ -30,10 +30,10 @@ sealed abstract class JCASigInterpreterPure[F[_], A](implicit M: Sync[F], signat
 
 }
 
-object JCASigInterpreterPure {
+object JCASigInterpreter {
 
-  def apply[F[_]: Sync, A: SigAlgoTag] = new JCASigInterpreterPure[F, A]() {}
+  def apply[F[_]: Sync, A: SigAlgoTag] = new JCASigInterpreter[F, A]() {}
 
-  implicit def genSig[F[_]: Sync, A: SigAlgoTag]: JCASigInterpreterPure[F, A] = apply[F, A]
+  implicit def genSig[F[_]: Sync, A: SigAlgoTag]: JCASigInterpreter[F, A] = apply[F, A]
 
 }
